Added omni shadow maps

+++ RTRenderer/DirectionalLight.cpp
@@ -15,8 +15,8 @@ DirectionalLight::DirectionalLight(Transform* transform)
 
 DirectionalLight::DirectionalLight(
 	Transform* transform,
-	GLfloat staticShadowWidth, GLfloat staticShadowHeight,
-	GLfloat dynamicShadowWidth, GLfloat dynamicShadowHeight,
+	GLuint staticShadowWidth, GLuint staticShadowHeight,
+	GLuint dynamicShadowWidth, GLuint dynamicShadowHeight,
 	GLfloat diffIntensity, GLfloat diffRed, GLfloat diffGreen, GLfloat diffBlue, 
 	GLfloat specIntensity, GLfloat specRed, GLfloat specGreen, GLfloat specBlue)
 	: Light(transform,
