+++ RTRenderer/Light.cpp
@@ -12,8 +12,8 @@ Light::Light(Transform* transform)
 }
 
 Light::Light(Transform * transform,
-	GLfloat staticShadowWidth, GLfloat staticShadowHeight,
-	GLfloat dynamicShadowWidth, GLfloat dynamicShadowHeight,
+	GLuint staticShadowWidth, GLuint staticShadowHeight,
+	GLuint dynamicShadowWidth, GLuint dynamicShadowHeight,
 	GLfloat diffIntensity, GLfloat diffRed, GLfloat diffGreen, GLfloat diffBlue, 
 	GLfloat specIntensity, GLfloat specRed, GLfloat specGreen, GLfloat specBlue)
 {
