+++ RTRenderer/Light.h
@@ -22,8 +22,8 @@ protected:
 public:
 	Light(Transform* transform);
 	Light(Transform* transform,
-		GLfloat staticShadowWidth, GLfloat staticShadowHeight,
-		GLfloat dynamicShadowWidth, GLfloat dynamicShadowHeight,
+		GLuint staticShadowWidth, GLuint staticShadowHeight,
+		GLuint dynamicShadowWidth, GLuint dynamicShadowHeight,
 		GLfloat diffIntensity, GLfloat diffRed, GLfloat diffGreen, GLfloat diffBlue, 
 		GLfloat specIntensity, GLfloat specRed, GLfloat specGreen, GLfloat specBlue);
